+++ printer/stub/xmake.lua
@@ -1,0 +1,10 @@
+add_rules("mode.debug", "mode.release")
+set_languages("c++20")
+
+target("printer-stub")
+    set_default(false)
+    set_kind("headeronly")
+    set_group("print")
+    add_includedirs("include", {public=true})
+    add_headerfiles("include/(**)")
+    add_deps("core")