+++ changelog.md
@@ -1,5 +1,11 @@
 Changelog
 =========
+2.1.0
+-----
+- Isolate Logger from LoggerInterface and Level
+- Created Printer::Stub()
+- Changed std::runtime_error to std::logic_error when Filter/Union receive nullptr
+
 2.0.1
 -----
 - Adjust header instalation